+++ src/components/ExersiceModalWindow/ExersiceModalWindow.module.css
@@ -65,6 +65,11 @@
   align-self: flex-start;
 }
 
+.ExersiceModalWindowBtn:hover {
+  background: var(--orange-1, #EF8964);
+  color: #efede8;
+}
+
 @media screen and (min-width: 768px) {
   .ExersiceModalWindow {
     display: flex;
