+++ src/components/ExersiceModalWindow/SuccessModalWindow/SuccessExerciseModalWindow/SuccessExerciseModalWindow.jsx
@@ -1,10 +1,10 @@
 import React, { useState } from "react";
-import css from "./AddProductSuccess.module.css";
+import BasicModalWindow from "../../../BasicModalWindow/BasicModalWindow";
+import css from "./SuccessModalWindow.module.css";
 import symbolDefs from "../../../assets/images/symbol-defs.svg";
-import avocado from "../../../assets/images/avocado.svg";
-import BasicModalWindow from "../../BasicModalWindow/BasicModalWindow";
+import thumb from "../../../../assets/images/thumb.svg";
 
-export const AddProductSuccess = () => {
+export const SuccessExerciseModalWindow = () => {
   const [isModalOpen, setIsModalOpen] = useState(false);
 
   const openModal = () => {
@@ -22,16 +22,20 @@ export const AddProductSuccess = () => {
             <div className={`${css.SuccessModalWindow} ${css.boxImage}`}>
               <img
                 className={css.SuccessModalWindowImg}
-                src={avocado}
-                alt="avocado"
+                src={thumb}
+                alt="thumb"
               />
               <p className={css.SuccessModalWindowTitle}>Well done</p>
               <p className={css.SuccessModalWindowTxt}>
-                Calories:{" "}
-                <span className={css.SuccessModalWindowSub}>96</span>
+                Your time:{" "}
+                <span className={css.SuccessModalWindowSub}> 3 minutes</span>
+              </p>
+              <p className={css.SuccessModalWindowTxt}>
+                Burned calories:{" "}
+                <span className={css.SuccessModalWindowSub}>150</span>
               </p>
             </div>
-            <button className={css.SuccessModalWindowBtn}>Next product</button>
+            <button className={css.SuccessModalWindowBtn}>Next exercise</button>
             <p className={css.SuccessModalWindowTxt}>
               To the diary
               <svg className={css.arrowIcon}>
