+++ src/components/ExersiceModalWindow/SuccessModalWindow/SuccessExerciseModalWindow/SuccessModalWindow.module.css
@@ -1,0 +1,78 @@
+.SuccessModalWindow {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.SuccessModalWindowWrap {
+  padding: 48px 89px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.SuccessModalWindowImg {
+  width: 118px;
+  height: 73px;
+  margin-bottom: 32px;
+}
+
+.SuccessModalWindowTitle {
+  color: #efede8;
+  font-size: 24px;
+  font-weight: 600;
+  line-height: 32px;
+  margin-bottom: 16px;
+}
+
+.SuccessModalWindowTxt {
+  color: rgba(239, 237, 232, 0.3);
+  font-size: 14px;
+  font-weight: 400;
+  line-height: 18px;
+  display: flex;
+  gap: 8px;
+  align-items: center;
+}
+
+.SuccessModalWindowSub {
+  color: #e6533c;
+  font-size: 14px;
+  font-weight: 400;
+  line-height: 18px;
+}
+
+.SuccessModalWindowBtn {
+  border-radius: 12px;
+  background: #e6533c;
+  padding: 14px 32px;
+  justify-content: center;
+  align-items: center;
+  margin-top: 24px;
+  margin-bottom: 16px;
+  color: #efede8;
+  font-size: 16px;
+  font-weight: 500;
+  line-height: 24px;
+}
+
+.SuccessModalWindowBtn:hover {
+  background: var(--orange-1, #EF8964);
+  color: #efede8;
+}
+
+.arrowIcon {
+  width: 16px;
+  height: 16px;
+  fill: var(--white);
+}
+
+@media screen and (min-width: 768px) {
+  .SuccessModalWindowWrap {
+    padding: 64px 134px;
+  }
+
+  .SuccessModalWindowBtn {
+    padding: 16px 32px;
+  }
+}
