+++ src/pages/Exercices/Exercices.jsx
@@ -3,7 +3,6 @@ import { TitlePage } from "../../components/TitlePage/TitlePage";
 import css from "./Exercices.module.css";
 import ExercisesCategories from "../../components/exercises/ExercisesCategories/ExercisesCategories";
 import { Outlet, useLocation } from "react-router-dom";
-import { SuccessModalWindow } from "../../components/ExersiceModalWindow/SuccessModalWindow/SuccessModalWindow";
 import Back from "@/components/exercises/Back/Back";
 
 const Exercises = () => {
@@ -19,7 +18,6 @@ const Exercises = () => {
         <header className={css.header}>
           <TitlePage text="Exercises" />
           <ExercisesCategories />
-          <SuccessModalWindow/>
         </header>
         <Outlet />
       </Container>
